chore: bootstrap PSR-4 autoloader
